+++ crates/oxc_minifier/tests/ecmascript/may_have_side_effects.rs
@@ -1183,6 +1183,8 @@ fn test_call_expressions() {
 
     test("URL.canParse()", false);
 
+    test("BigInt64Array.of()", false);
+    test("BigUint64Array.of()", false);
     test("Float32Array.of()", false);
     test("Float64Array.of()", false);
     test("Int16Array.of()", false);
